+++ src/api/resources/wallets.js
@@ -5,9 +5,12 @@ import { Keypair } from '../../base/keypair'
 import { makeChangeSignerTransaction } from './change_signers'
 import * as errors from '../../errors'
 
+const HORIZON_VERSION_PREFIX = 'v3'
+
 /**
  * Wallets.
  */
+
 export class Wallets extends ResourceGroupBase {
   /**
    * Get key derivation params.
@@ -260,6 +263,7 @@ export class Wallets extends ResourceGroupBase {
    * Change password.
    *
    * @param {string} newPassword Desired password.
+   *
    * @return {Wallet} New wallet.
    */
   async changePassword (newPassword) {
@@ -339,8 +343,14 @@ export class Wallets extends ResourceGroupBase {
   }
 
   _getSigners (accountId) {
-    return this._sdk.horizon.account.getSigners(accountId)
-      .then(response => response.data.signers)
+    return this._server
+      ._makeCallBuilder()
+      .appendUrlSegment(HORIZON_VERSION_PREFIX)
+      .appendUrlSegment('accounts')
+      .appendAccountId(accountId)
+      .appendUrlSegment('signers')
+      .get()
+      .then(response => response.data)
       .catch(err => {
         if (err instanceof errors.NotFoundError) {
           return []