+++ src/index.js
@@ -9,7 +9,7 @@ export { Wallet } from './wallet'
 export { HorizonResponse } from './horizon'
 export { ApiResponse } from './api'
 
-export { ApiCaller } from './api2'
+export { ApiCaller, JsonapiResponse } from './api2'
 export { KeyServerCaller } from './api2/key-server-caller'
 
 // Typed errors